Refactor the fragments
