+++ app/build.gradle
@@ -57,6 +57,4 @@ dependencies {
     implementation 'com.squareup.retrofit2:retrofit:2.7.2'
     implementation 'com.squareup.retrofit2:converter-gson:2.7.2'
     implementation 'com.squareup.okhttp3:okhttp:3.14.7'
-
-    implementation "android.arch.lifecycle:extensions:1.1.1"
 }
