+++ app/src/main/java/com/everest/movieapp/adapters/SectionsPagerAdapter.kt
@@ -5,8 +5,8 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentPagerAdapter
 import com.everest.movieapp.R
-import com.everest.movieapp.ui.main.fragments.CurrentYearMovies
-import com.everest.movieapp.ui.main.fragments.PopularMovies
+import com.everest.movieapp.ui.CurrentYearMovies
+import com.everest.movieapp.ui.PopularMovies
 
 private val TAB_TITLES = arrayOf(
     R.string.tab_text_1,
