+++ app/src/main/java/com/everest/movieapp/ui/PopularMovies.kt
@@ -1,22 +1,24 @@
-package com.everest.movieapp.ui.main.fragments
+package com.everest.movieapp.ui
 
 import android.os.Bundle
-import android.view.*
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.RecyclerView
 import com.everest.movieapp.R
+import com.everest.movieapp.adapters.MovieRecyclerViewAdapter
 import com.everest.movieapp.api.Api
 import com.everest.movieapp.api.RetrofitHelper
 import com.everest.movieapp.databinding.FragmentPopularMoviesBinding
 import com.everest.movieapp.model.MovieDb
-import com.everest.movieapp.adapters.MovieRecyclerViewAdapter
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
-class PopularMovies:Fragment(R.layout.fragment_current_year_movies) {
+class PopularMovies : Fragment(R.layout.fragment_current_year_movies) {
 
-    private lateinit var binding : FragmentPopularMoviesBinding
+    private lateinit var binding: FragmentPopularMoviesBinding
     private lateinit var recyclerView: RecyclerView
     private lateinit var movieRecyclerViewAdapter: MovieRecyclerViewAdapter
 
@@ -25,9 +27,9 @@ class PopularMovies:Fragment(R.layout.fragment_current_year_movies) {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentPopularMoviesBinding.inflate(layoutInflater,container,false)
+        binding = FragmentPopularMoviesBinding.inflate(layoutInflater, container, false)
 
-        recyclerView= binding.popularRecyclerView
+        recyclerView = binding.popularRecyclerView
 
         populateData()
