+++ app/src/main/res/layout/fragment_current_year_movies.xml
@@ -7,7 +7,7 @@
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.PlaceholderFragment">
+    tools:context=".ui.PlaceholderFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/current_year_recycler_view"
