+++ app/src/main/res/layout/fragment_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:backgroundTint="@color/black"
-    tools:context=".ui.main.PlaceholderFragment">
+    tools:context=".ui.PlaceholderFragment">
 
     <TextView
         android:id="@+id/section_label"
