+++ app/src/main/res/layout/fragment_popular_movies.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:backgroundTint="@color/black"
-    tools:context=".ui.main.PlaceholderFragment">
+    tools:context=".ui.PlaceholderFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/popular_recycler_view"
